add toc drawer test
